Rename strapi-dbal to strapi-database
