+++ packages/strapi-database/LICENSE.md
