+++ packages/strapi-database/README.md
@@ -1,9 +1,9 @@
-# Strapi database abstraction layer
+# Strapi database layer
 
-[![npm version](https://img.shields.io/npm/v/strapi-dbal.svg)](https://www.npmjs.org/package/strapi-dbal)
-[![npm downloads](https://img.shields.io/npm/dm/strapi-dbal.svg)](https://www.npmjs.org/package/strapi-dbal)
-[![npm dependencies](https://david-dm.org/strapi/strapi-dbal.svg)](https://david-dm.org/strapi/strapi-dbal)
-[![Build status](https://travis-ci.org/strapi/strapi-dbal.svg?branch=master)](https://travis-ci.org/strapi/strapi-dbal)
+[![npm version](https://img.shields.io/npm/v/strapi-database.svg)](https://www.npmjs.org/package/strapi-database)
+[![npm downloads](https://img.shields.io/npm/dm/strapi-database.svg)](https://www.npmjs.org/package/strapi-database)
+[![npm dependencies](https://david-dm.org/strapi/strapi-database.svg)](https://david-dm.org/strapi/strapi-database)
+[![Build status](https://travis-ci.org/strapi/strapi-database.svg?branch=master)](https://travis-ci.org/strapi/strapi-database)
 [![Slack status](https://slack.strapi.io/badge.svg)](http://slack.strapi.io)
 
 This package is strapi's database handling layer. It is responsible for orchestrating database connectors and implementing the logic of strapi's data structures.
