+++ packages/strapi-database/lib/database-manager.js
