+++ packages/strapi-database/lib/naming.js
