+++ packages/strapi-database/lib/require-connector.js
