+++ packages/strapi-database/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "strapi-dbal",
+  "name": "strapi-database",
   "version": "3.0.0-beta.17.4",
-  "description": "Strapi's database abstraction layer",
+  "description": "Strapi's database layer",
   "homepage": "http://strapi.io",
   "main": "./lib/index.js",
   "scripts": {
