+++ packages/strapi/package.json
@@ -57,7 +57,7 @@
     "strapi-generate-policy": "3.0.0-beta.17.4",
     "strapi-generate-service": "3.0.0-beta.17.4",
     "strapi-utils": "3.0.0-beta.17.4",
-    "strapi-dbal": "3.0.0-beta.17.4"
+    "strapi-database": "3.0.0-beta.17.4"
   },
   "scripts": {
     "test": "jest --verbose",
